pyproject.toml
[tool.poetry]
name = "taptrust"
version = "0.1.0"
description = "Tap-to-verify mobile driver’s licenses (ISO 18013-5) in Python"
authors = ["TapTrust Team <team@taptrust.org>"]
license = "MIT"

# Tell Poetry “Hey, our Python package is under the `verifier/` directory.”
packages = [
  { include = "verifier" }
]

[tool.poetry.scripts]
idscan = "verifier.cli:app"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
cbor2 = "*"
pycose = "*"
cryptography = "*"
typer = "*"      # ← add this line
nfcpy = "*"      # ← add this line
oscrypto = "^1.3.0"
PyYAML = "*"
pyscard = "*"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4"
flake8 = "^6.0.0"
pre-commit = "^3.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
